Setup i18n & Notify

+++ src/app/modules/carts/components/Item.js
@@ -1,11 +1,14 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
+import { toast } from 'react-toastify';
+import { useTranslation } from "react-i18next";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import { actionDeleteProduct } from '../actions';
 
 const CartItem = ({ cart, deleteProduct }) => {
+  const { t } = useTranslation();
   const { id, name, price, image_url } = cart.product;
 
   const countTotal = (price, quantity) => {
@@ -13,7 +16,11 @@ const CartItem = ({ cart, deleteProduct }) => {
   }
 
   const handleDelete = cart => {
-    deleteProduct(cart);
+    const confirm = window.confirm(t('NOTIFY.DELETE_CART_CONFIRM', { name: cart.name }));
+    if (confirm) {
+      deleteProduct(cart);
+      toast.error(t('NOTIFY.DELETE_PRODUCT_SUCCESS', { name: cart.name }));
+    }
   }
 
   return (
